RF | Combining FolderTest classes
